requirements.txt
streamlit
pandas
dash
python-dotenv
"snowflake-connector-python[secure-local-storage,pandas]"
